Rename files and fix EOL
